Bugfix index.html: Load sync instead of async
There was a bug that showed up in the menu.
Sometimes the labels would load and sometimes not.
The reason was that elements were loaded asynchronously in loadFragment.
That turned into a race-condition sometimes.
In those cases the i18n function would run too early,
at a point when the elements were not available yet.