New French translations
As always, I am not a native speaker so it would be best to find a native speaker to take over (and correct my translations).

One thing I didn't know how to account for: traduction and dictionnaire have different genders. So "est déjà installé" would change to "est déjà installée" depending on the gender, for example.